test: cleanup box/net.box_call_blocks_gh-946
Found issue with 'readahead limit is reached':

 iproto iproto.cc:606 W> stopping input on connection fd 23, aka unix/:(socket), peer of unix/:(socket), readahead limit is reached

While the test doesn't check it. To fix the issue increased
'readahead' value.

Found issue at the tests running after the current test:

[001] --- box/net.box_disconnect_gh-3859.result	Tue Jun 30 19:16:28 2020
[001] +++ box/net.box_disconnect_gh-3859.reject	Fri Jul  3 17:38:22 2020
[001] @@ -97,7 +97,7 @@
[001]  ...
[001]  disconnected -- nil, on_disconnect is not called second time.
[001]  ---
[001] -- null
[001] +- true
[001]  ...
[001]  box.session.on_disconnect(nil, on_disconnect)
[001]  ---
[001]

 main/435/console/unix/: utils.c:1005 E> LuajitError: [string "function on_disconnect() disconnected = true ..."]:1: assign to undeclared variable 'disconnected'

Found that the root cause of the issue, that tests run on the same
worker, in this case function on_disconnected must be cleaned.
Also clean up of the test at its end improved.

Closes #5133